FLAC codec: fixed 64bit issue
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@2008 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ plugins/codec_flac/FlacDecoder.h
@@ -101,7 +101,7 @@ protected:
      * @return read state
      */
     virtual ::FLAC__StreamDecoderReadStatus read_callback(
-        FLAC__byte buffer[], unsigned int *bytes);
+        FLAC__byte buffer[], size_t *bytes);
 
     /**
      * FLAC decoder interface: write callback.
